Cleaned up the scene and scene loading system
Made an engine package for all of the entity and scene related stuff. Renamed the ResourceLoader to SceneLoader. Made pre and post load events for the game-side handler so it can create an actual Terrain entity, send it back to the engine for mesh loading, then post process it. Otherwise it was just a straight up Entity.

Now supporting the idea of non-renderable entities that can still be part of the scene, as well as child entities.

Experimented with loading .blend files, but the custom properties are not stored there. For FBX that leaves me with the choice of using FBX ASCII which doesn't support custom properties, or FBX Binay which doesn't support materials... materials lose.