Ui scaling chainge, and add support for xMax and yMax relative anchors
Ui scaling is similar to Unity now. Canvas immediately takes on the working resolution rather than the reference resolution so it always fills the whole window. UiElement scaling is recalculating individual parts rather than using a post process scale so that relative anchors can be calculated properly for the next feature below...

When x anchors or y anchors deviate in value, the UiElement's xMax or yMax respectively switch from representing width / height to representing an offset from the max anchor.

As far as I can tell, this seems to match Unity's behavior, which allows fully fluid resolution / window sizes if the developer chooses to base UI elements on anchors rather than reference resolution. Either is valid, but the new method allows buttons etc to be pinned to the corners or edges of the screen quite nicely.