new scene input and update system
scene has an empty root entity. any entity parented to it will be added to the scene for input processing and updates - including the entities entire ancestory if it is not the root.

The game has to call Scene.input and Scene.update to trigger the process - similar to SceneRenderer.render. The UiElements gate update mesh rebuilding with a dirty flag, and a recursive updateSize, or a constructor call will set the flag, causing the next auto update to build the mesh and add it to the scene for rendering.