Simplifying core engine services
Trying out some DIY "dependency injection" technique. Deleted the Engine class since it was pretty much a globals shed. Moved its functionality into Main where the game can new each service it needs and pass it to the Game. Moved the Scene construction out to Main, and it gets passed to Game. Game creates the camera, attaches it to the scene, and no longer has to manually update it. Game no longer has to manage SceneManager directly - and scene manager gets its camera from Scene instead of as an argument.