Big shader and SceneRenderer refactor
Molding SceneRenderer into something more to my tastes. Instead of specific types of objects hard coded in (skybox, GUI elements), the shaders are set dynamically on materials. All meshes added to the scene are mapped according to their shader. The shaders that have been mapped are iterated, their mesh lists rendered once per entity.

Scene is updated so that entities with no meshes can be added, and when an entity update causes a mesh to be generated or modified, the scene will slurp it up and add it to the shader to mesh mapping, and the mesh to entity mapping.

IHud garbage has been removed. The SceneRenderer is now only explicit about lights, which seems totally fine to me.

This change caused what appeared to be a pretty major framerate drop, but it was actually related to the framerate text mesh not refreshing correctly and getting clipped from 2k+ to 200+. It was a really PITA to track down.