2nd pass on basic collision
This system takes velocity and clips it based on collisions, then applies the final resolved velocity at the end of the frame.

It also handles both bodies being dynamic, where it scales the resolution vectors by 50%. It also handles one to many collisions in one frame by averaging the resolution vectors and combining them once after all the collisions have been detected.

It does not handle one collision resolution resulting in another collision, or one to many collisions where the many are in opposing directions.

So, it's not an open world physics sim, but sufficient for basic games where there are multiple actors and an environment.